pyproject.toml
[tool.poetry]
name = "steward-discord"
version = "0.1.0"
description = "Discord part of steward"
authors = ["AulonSal <aulonsal@gmail.com>"]
license = "UNLICENSE"

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
